(svn r18996) -Fix (r18993, r18994): MSVC 64 bits had somethings to complain about

+++ src/core/smallvec_type.hpp
@@ -79,9 +79,9 @@ public:
 	 * @param to_add the number of items to append
 	 * @return pointer to newly allocated item
 	 */
-	FORCEINLINE T *Append(size_t to_add = 1)
+	FORCEINLINE T *Append(uint to_add = 1)
 	{
-		size_t begin = this->items;
+		uint begin = this->items;
 		this->items += to_add;
 
 		if (this->items > this->capacity) {
